+++ src/constans/Colors.ts
@@ -3,4 +3,5 @@ export default {
     black: '#000',
     primary: "#4DB7AD",
     secondary: "#b1b8b0",
+    tertiary: "#7DA1BF",
 };